[ENH] add base annotator to make it easier to inherit
