+++ src/alpaca_eval/annotators/pairwise_evaluator.py
@@ -1,24 +1,26 @@
 import logging
-import os
-from functools import partial
 from pathlib import Path
-from typing import Any, Callable, Optional, Sequence, Type, Union
+from typing import Any, Callable, Optional, Sequence, Union
 
 import numpy as np
 import pandas as pd
 
-from .. import completion_parsers, constants, utils
-from ..decoders import get_fn_completions
-
-CURRENT_DIR = Path(__file__).parent
-logging.getLogger().setLevel(logging.INFO)
+from .. import utils
+from .base import BaseAnnotatorJSON, SingleAnnotator
 
 __all__ = ["PairwiseAnnotator", "SinglePairwiseAnnotator"]
 
 
-class PairwiseAnnotator:
-    """Class for a pool of annotators.
-
+class PairwiseAnnotator(BaseAnnotatorJSON):
+    __doc__ = (
+        BaseAnnotatorJSON.__doc__.replace("Base class", "Class")
+        + """
+    
+    p_label_flip : float, optional
+        Probability of flipping the label (ie adds noise by taking a mixture between predicted label and
+        2*p_label_flip of independent coin flip). If None, will not flip the label. In AlpacaFarm we use 0.25
+        for training. You can set this later on using `set_noise`.
+        
     Notes
     -----
     There are three main functions for annotations depending on how the outputs to compare are given:
@@ -27,123 +29,27 @@ class PairwiseAnnotator:
             outputs. Useful for collecting pairwise preferences for RLHF.
         - annotate_head2head: annotate a pair of sequence of outputs, each containing `"output"` which will be merged
             into a single sequence of paired outputs. Useful for evaluation against a reference.
-
-    Other functions that are useful for annotating:
-        - set_noise: set the noise level for the annotators.
-        - load_: load annotations from a file.
-        - save: save annotations to a file.
-
-    Parameters
-    ----------
-    annotators_config : Path or list of dict, optional
-        A dictionary or path to a yaml file containing the configuration for the pool of annotators. If a directory,
-        we search for 'configs.yaml' in it. The keys in the first  dictionary should be the annotator's name, and
-        the value should be a dictionary of the annotator's configuration which should have the following keys:
-        The path is relative to `evaluators_configs/` directory.
-        - prompt_template (str): a prompt template or path to it. The template should contain placeholders for keys in
-            the example dictionary, typically {instruction} and {output_1} {output_2}.
-        - fn_completions (str): function in `alpaca_farm.decoders` for completions. Needs to accept as first argument
-            `prompts` which is a list of string.
-        - completions_kwargs (dict): kwargs for fn_completions. E.g. model_name, max_tokens, temperature,
-        tokens_to_avoid
-        - fn_completion_parser (str) : Function in `completion_parsers.py` to use for parsing the completions into
-        preferences.
-        - completion_parser_kwargs (dict) : Kwargs for fn_completion_parser.
-        - other kwargs to `SinglePairwiseAnnotator` such as batch_size
-
-    seed : int, optional
-        Seed for the random number generator.
-
-    is_avoid_reannotations : bool, optional
-        Whether to avoid re-annotating examples that have already been annotated by the annotator. This will decrease
-        cost but can be slightly slower if there are no annotations that can be reused.
-
-    caching_path : Path, optional
-        Path to cache the annotations to. If None, will not save the annotations. If the path already exists it will
-        load annotations from there.
-
-    input_keys : tuple of str, optional
-        Keys use to distinguish inputs.
-
-    output_keys : tuple of str, optional
-        Keys use to distinguish outputs.
-
-    p_label_flip : float, optional
-        Probability of flipping the label (ie adds noise by taking a mixture between predicted label and
-        2*p_label_flip of independent coin flip). If None, will not flip the label. In AlpacaFarm we use 0.25
-        for training. You can set this later on using `set_noise`.
-
-    other_keys_to_keep : tuple of str, optional
-        Other columns to store besides the preferences.
-
-    is_store_missing_preferences : bool, optional
-        Whether to store missing preferences. If True it will avoid trying to always reannotate examples that have
-        errors.
-
-    base_dir : Path, optional
-        Path to the directory containing the annotators configs. I.e. annotators_config will be relative
-        to this directory.
     """
+    )
 
     def __init__(
         self,
-        annotators_config: Union[utils.AnyPath, list[dict[str, Any]]] = "claude",
-        seed: Optional[int] = 0,
-        is_avoid_reannotations: bool = True,
-        caching_path: Optional[utils.AnyPath] = "auto",
+        *args,
         input_keys: Sequence[str] = ("instruction",),
         output_keys: Sequence[str] = ("output_1", "output_2"),
         p_label_flip: Optional[float] = None,
-        other_keys_to_keep: Sequence[str] = ("price_per_example", "time_per_example"),
-        is_store_missing_preferences: bool = True,
-        base_dir: utils.AnyPath = constants.EVALUATORS_CONFIG_DIR,
+        **kwargs,
     ):
-        logging.info(f"Creating the annotator from `{annotators_config}`.")
-        self.base_dir = Path(base_dir)
-
-        # setting it relative to the config directory
-        annotators_config = self.base_dir / annotators_config
-
-        if annotators_config.is_dir():
-            annotators_config = annotators_config / "configs.yaml"
-
-        if caching_path == "auto":
-            if isinstance(annotators_config, (str, Path, os.PathLike)):
-                stem = Path(annotators_config).stem
-                caching_path = Path(annotators_config).parent / f"annotations_seed{seed}_{stem}.json"
-                logging.info(f"Saving annotations to `{caching_path}`.")
-            else:
-                logging.warning("caching_path cannot be 'auto' if annotators_config is not a path. Setting to None.")
-                caching_path = None
-        elif caching_path is not None:
-            logging.warning("Saving_path is given but not 'auto', make sure that it's different for different seeds.")
-
-        self.seed = seed
-        self.is_avoid_reannotations = is_avoid_reannotations
-        self.input_keys = list(input_keys)
-        self.output_keys = list(output_keys)
-        self.input_output_keys = self.input_keys + self.output_keys
-        self.all_keys = self.input_keys + self.output_keys + ["annotator"]
-        self.other_keys_to_keep = list(other_keys_to_keep)
+        super().__init__(*args, **kwargs, input_keys=input_keys, output_keys=output_keys)
         self.p_label_flip = p_label_flip
-        self.is_store_missing_preferences = is_store_missing_preferences
-        self.annotators_config = annotators_config
 
-        self.annotators = self._initialize_annotators(annotators_config)
-        self.caching_path = caching_path
-        self.df_annotations = None
-        self.reinitialize_cache_()
-
-    ### Helper properties to make it easier to inherit from this class ###
     @property
-    def SingleAnnotator(self):
+    def SingleAnnotator(self) -> SingleAnnotator:
         return SinglePairwiseAnnotator
 
-    #########################################
-
     @property
-    def annotator_name(self) -> str:
-        return Path(self.annotators_config).parent.name
+    def annotation_key(self) -> str:
+        return "preference"
 
     def annotate_samples(
         self,
@@ -226,7 +132,7 @@ class PairwiseAnnotator:
             self.set_noise(p_label_flip)
 
         try:
-            annotated = self.annotate_pairs(df_to_annotate, **decoding_kwargs)
+            annotated = self.__call__(df_to_annotate, **decoding_kwargs)
         finally:
             # reset even if there is an error
             if p_label_flip is not None:
@@ -317,7 +223,7 @@ class PairwiseAnnotator:
                     """
                 )
 
-        out = self.annotate_pairs(df_to_annotate, **decoding_kwargs)
+        out = self.__call__(df_to_annotate, **decoding_kwargs)
 
         return out
 
@@ -342,14 +248,8 @@ class PairwiseAnnotator:
             The annotated examples. Each dictionary will contain all of `self.input_output_keys` and `"preference"`.
             Preference will be 0 if output_1 == output_2, 1 if output_1 is preferred, and 2 if output_2 is preferred.
         """
-        if len(to_annotate) == 0:
-            return []
-
-        df_to_annotate = self._preprocess(to_annotate)
-
-        df_annotated = self._annotate(df_to_annotate, **decoding_kwargs)
-        annotated = self._postprocess_and_store_(df_annotated, to_annotate)
-        return annotated
+        # `annotate_pairs` is used for backward compatibility
+        return self.__call__(to_annotate, **decoding_kwargs)
 
     def set_noise(self, p_label_flip: float):
         """Set the noise level for the annotators.
@@ -364,32 +264,11 @@ class PairwiseAnnotator:
         self.p_label_flip = p_label_flip
 
     def _preprocess(self, to_annotate: utils.AnyData) -> pd.DataFrame:
-        """Preprocess the examples to annotate. In particular takes care of filtering unnecessary examples."""
+        # same as preprocess but with potential random noising and dealing with eauality
 
-        df_to_annotate = utils.convert_to_dataframe(to_annotate).copy()
+        df_to_annotate = super()._preprocess(to_annotate)
 
-        for c in self.other_keys_to_keep + ["preference"]:
-            if c in df_to_annotate.columns:
-                logging.warning(f"""{c} column is already in the dataframe. We will overwrite it.""")
-                df_to_annotate[c] = np.nan
-
-        # remove duplicates because you only need to annotate one of them
-        df_to_annotate = df_to_annotate.drop_duplicates(subset=self.input_output_keys)
-
-        # set the annotater for each example
-        df_to_annotate["annotator"] = df_to_annotate.apply(
-            lambda x: utils.random_seeded_choice(
-                # we add "annotator" at the beginning to not use the same seed for all tasks
-                seed="annotator" + x["instruction"] + str(self.seed),
-                choices=list(self.annotators.keys()),
-            ),
-            axis=1,
-        )
-
-        if self.is_avoid_reannotations:
-            df_to_annotate = self.apply_cached_annotations(df_to_annotate)
-
-        # adds random noise => avoids annotating examples that will be noised out.
+        # 1. adds random noise => avoids annotating examples that will be noised out.
         if self.p_label_flip:
             logging.info(f"Adding random noise to the labels p_label_flip={self.p_label_flip}.")
             # if you have 25% change of flipping the label, you have 50% chance of selecting random label
@@ -405,308 +284,54 @@ class PairwiseAnnotator:
             )
             df_to_annotate["is_noisy_label"] = ~noisy_preference.isna()
             # keeps previously annotated examples when you did not add noise
-            df_to_annotate["preference"] = np.where(
+            df_to_annotate[self.annotation_key] = np.where(
                 df_to_annotate["is_noisy_label"],
                 noisy_preference,
-                df_to_annotate["preference"],
+                df_to_annotate[self.annotation_key],
             )
 
+        # 2. deals with equality
         idcs_is_same_outputs = df_to_annotate["output_1"] == df_to_annotate["output_2"]
-        df_to_annotate.loc[idcs_is_same_outputs, "preference"] = 0
+        df_to_annotate.loc[idcs_is_same_outputs, self.annotation_key] = 0
 
         return df_to_annotate
 
-    def apply_cached_annotations(self, df_to_annotate: pd.DataFrame) -> pd.DataFrame:
-        """annotate examples with cached annotations"""
-        df_to_annotate = self._merge_annotations(df_to_annotate, self.df_annotations)
-        return df_to_annotate
-
-    def _initialize_annotators(
-        self, annotators_config: Union[utils.AnyPath, dict[str, dict[str, Type["SinglePairwiseAnnotator"]]]]
-    ) -> dict[str, Callable]:
-        """Load all the configs and prompts if necessary."""
-        annotators_config = utils.load_configs(annotators_config)
-        return {
-            name: self.SingleAnnotator(seed=self.seed, base_dir=self.base_dir, **annotator_config)
-            for name, annotator_config in annotators_config.items()
-        }
-
-    def _annotate(self, df_to_annotate: pd.DataFrame, **decoding_kwargs) -> pd.DataFrame:
-        """Annotate the examples."""
-
-        df_annotated = df_to_annotate
-        for annotator in self.annotators.keys():
-            # only annotate examples that have not been annotated yet
-            curr_idcs = (df_annotated["annotator"] == annotator) & df_annotated["preference"].isna()
-
-            logging.info(f"Annotating {curr_idcs.sum()} examples with {annotator}")
-
-            # actual annotation
-            curr_annotated = self.annotators[annotator](df_annotated.loc[curr_idcs, self.all_keys], **decoding_kwargs)
-
-            df_annotated = self._merge_annotations(df_annotated, curr_annotated)
-
-        return df_annotated
-
-    def _postprocess_and_store_(
-        self,
-        df_annotated: pd.DataFrame,
-        to_annotate: Union[Sequence[dict[str, Any]], pd.DataFrame],
-    ) -> list[dict[str, Any]]:
-        """Convert the dataframe into a list of dictionaries to be returned, and store current anntations."""
-
-        df_to_annotate = utils.convert_to_dataframe(to_annotate)
-
-        # select available annotations
-        if self.is_store_missing_preferences:
-            df_annotated["preference"] = df_annotated["preference"].fillna(-1)
-        else:
-            df_annotated["preference"] = df_annotated["preference"].replace(-1, np.nan)
-
-        df_annotated = df_annotated[~df_annotated["preference"].isna()].copy()
-
-        # try converting to int now that no nan
-        df_annotated["preference"] = pd.to_numeric(df_annotated["preference"], downcast="integer", errors="ignore")
-
+    def _store_annotations_(self, df_annotated: pd.DataFrame):
         if "is_noisy_label" in df_annotated.columns:
-            # dont' store noisy labels
-            df_annotated_to_store = df_annotated.query("is_noisy_label == False").drop(columns=["is_noisy_label"])
-            df_annotated = df_annotated.drop(columns=["is_noisy_label"])
-        else:
-            df_annotated_to_store = df_annotated
+            # don't store noisy labels
+            df_annotated = df_annotated.query("is_noisy_label == False").drop(columns=["is_noisy_label"])
 
-        other_keys_to_keep = [c for c in self.other_keys_to_keep if c in df_annotated_to_store.columns]
-        all_keys_to_keep = self.all_keys + ["preference"] + other_keys_to_keep
-        df_annotated_to_store = df_annotated_to_store[all_keys_to_keep]
-
-        self.store_annotations_(df_annotated_to_store)
-
-        if self.is_store_missing_preferences:
-            # put back np.nan
-            df_annotated["preference"] = df_annotated["preference"].replace(-1, np.nan)
-
-        # need to merge with df_to_annotate in case you dropped duplicates
-        on = list(self.input_keys + self.output_keys)
-        df_annotated = df_annotated[all_keys_to_keep]
-        df_to_annotate = df_to_annotate[[c for c in df_to_annotate.columns if c not in df_annotated.columns or c in on]]
-        # need to remove all other columns before merging if not you will
-        df_annotated = df_to_annotate.merge(df_annotated, on=on, how="outer")
-
-        annotated = df_annotated.to_dict(orient="records")
-
-        return annotated
-
-    def store_annotations_(self, df_annotated_to_store: pd.DataFrame):
-        """Store annotation in memory and on disk"""
-        if self.df_annotations is None:
-            df_annotations = df_annotated_to_store
-        else:
-            df_annotations = pd.concat([self.df_annotations, df_annotated_to_store], axis=0, ignore_index=True)
-
-        self.df_annotations = df_annotations.drop_duplicates(subset=self.all_keys, keep="last")
+        super()._store_annotations_(df_annotated)
 
         self.save()
 
-    def save(self, path: Optional[utils.AnyPath] = None):
-        """Save the annotations to json."""
-        path = path or self.caching_path
-        if path is not None:
-            logging.info(f"Saving all annotations to {path}.")
-            # to make sure that we don't overwrite the annotations we load again from file (ideally would use a DB)
-            self._refresh_annotations_()
-            if not self.is_store_missing_preferences:
-                self.df_annotations = self.df_annotations[~self.df_annotations["preference"].isna()]
-            self.df_annotations.to_json(path, orient="records", indent=2)
 
-    def _refresh_annotations_(self):
-        """Refresh the annotations in memory."""
-        curr_df_annotations = self.df_annotations.copy()
-        self.load_()
-        self.df_annotations = pd.concat(
-            [self.df_annotations, curr_df_annotations], axis=0, ignore_index=True
-        ).drop_duplicates(subset=self.all_keys, keep="last")
-
-    def reinitialize_cache_(self):
-        self.load_()
-
-    def load_(self, path: Optional[utils.AnyPath] = None):
-        """Load all the annotations from json."""
-        path = path or self.caching_path
-        if path is not None:
-            path = Path(path)
-            if path.exists():
-                logging.info(f"Loading all annotations from {path}.")
-                self.df_annotations = pd.read_json(path, dtype={k: str for k in self.all_keys})
-
-    def _merge_annotations(self, df_to_annotate: pd.DataFrame, df_partially_annotated: pd.DataFrame) -> pd.DataFrame:
-        """Merge (partial) annotations with the original df to keep the same order and avoid duplicates annotations."""
-
-        if df_partially_annotated is None or df_partially_annotated.empty:
-            return df_to_annotate
-
-        other_keys_to_keep = [c for c in self.other_keys_to_keep if c in df_partially_annotated.columns]
-
-        kwargs = dict(
-            on=self.all_keys,
-            how="left",
-            suffixes=("_old", "_new"),
+class SinglePairwiseAnnotator(SingleAnnotator):
+    __doc__ = (
+        SingleAnnotator.__doc__.replace(
+            "A helper class for a single auto annotator.",
+            "A helper class for a single pairwise auto annotator.",
         )
-        try:
-            df_to_annotate = df_to_annotate.merge(
-                df_partially_annotated[self.all_keys + ["preference"] + other_keys_to_keep], **kwargs
-            )
-        except ValueError:
-            # can have merging issues if columns have different dtypes
-            df_partially_annotated = df_partially_annotated.astype({k: str for k in self.all_keys})
-            df_to_annotate = df_to_annotate.astype({k: str for k in self.all_keys}).merge(
-                df_partially_annotated[self.all_keys + ["preference"] + other_keys_to_keep], **kwargs
-            )
-
-        # if columns were in both dataframes, try to merge them
-        for c in other_keys_to_keep + ["preference"]:
-            if f"{c}_old" in df_to_annotate.columns and f"{c}_new" in df_to_annotate.columns:
-                df_to_annotate[c] = df_to_annotate[c + "_old"].fillna(df_to_annotate[c + "_new"])
-                df_to_annotate = df_to_annotate.drop(columns=[c + "_old", c + "_new"])
-
-        return df_to_annotate
-
-
-class SinglePairwiseAnnotator:
-    """A helper class for a single auto annotators.
-
-    Parameters
-    ----------
-    prompt_template : str or path
-        A prompt template that will be given to `fn_prompter` or path to those prompts. Path is relative to
-        `evaluators_configs/`
-
-    fn_completion_parser : callable or str
-        Function in `completion_parsers.py` to use for parsing the completions into preferences. For each completion,
-        the number of preferences should be equal to the batch_size if not we set all the preferences in that batch to
-        NaN.
-
-    completion_parser_kwargs : dict
-        Kwargs for fn_completion_parser.
-
-    fn_completions : callable or str
-        Function in `decoders.py` to use for decoding the output.
-
-    completions_kwargs : dict
-        kwargs for fn_completions. E.g. model_name, max_tokens, temperature, top_p, top_k, stop_seq.
-
+        + """
     is_randomize_output_order : bool
         Whether to randomize output_1, output_2 when formatting.
-
-    is_shuffle : bool
-        Whether to shuffle the order of the examples before making the prompt. Useful if batch_size > 1.
-
-    seed : int
-        Seed for randomization.
-
-    batch_size : int
-        Number of examples that will be added in a single prompt.
-
-    base_dir : Path, optional
-        Path to the directory containing the annotators configs. I.e. annotators_config will be relative
-        to this directory.
     """
+    )
 
     def __init__(
         self,
-        prompt_template: utils.AnyPath,
-        fn_completion_parser: Union[Callable, str] = "regex_parser",
-        completion_parser_kwargs: Optional[dict[str, Any]] = None,
-        fn_completions: Union[Callable, str] = "openai_completions",
-        completions_kwargs: Optional[dict[str, Any]] = None,
+        *args,
         is_randomize_output_order: bool = True,
-        is_shuffle: bool = True,
-        seed: Optional[int] = 123,
-        batch_size: int = 1,
-        base_dir: utils.AnyPath = constants.EVALUATORS_CONFIG_DIR,
+        **kwargs,
     ):
-        self.base_dir = Path(base_dir)
-        self.prompt_template = self._get_prompt_template(prompt_template)
-
-        if isinstance(fn_completion_parser, str):
-            fn_completion_parser = getattr(completion_parsers, fn_completion_parser)
-        completion_parser_kwargs = completion_parser_kwargs or {}
-        self.fn_completion_parser = partial(fn_completion_parser, **completion_parser_kwargs)
-
+        super().__init__(*args, **kwargs)
         self.is_randomize_output_order = is_randomize_output_order
-        self.fn_completions = get_fn_completions(fn_completions)
-        self.completions_kwargs = completions_kwargs or {}
-        self.seed = seed
-        self.is_shuffle = is_shuffle
-        self.batch_size = batch_size
 
-    def _get_prompt_template(self, prompt_template: utils.AnyPath):
-        return utils.read_or_return(self.base_dir / prompt_template)
+    @property
+    def annotation_column(self):
+        return "preference"
 
-    def make_prompts(
-        self, df_to_annotate: pd.DataFrame, prompt_template: Optional[str] = None
-    ) -> tuple[list[str], pd.DataFrame]:
-        """Make all the prompts for the given examples.
-
-        Parameters
-        ----------
-        df_to_annotate : pd.DataFrame
-            Examples to annotate
-
-        prompt_template : str
-            Template to use for the prompt. If None, use the one from the constructor.
-
-        Returns
-        -------
-        prompts : list[str]
-            Formatted prompts for the given examples.
-
-        df_to_annotate : pd.DataFrame
-            Examples to annotate in the same order as prompts.
-        """
-        if prompt_template is None:
-            prompt_template = self.prompt_template
-        return utils.make_prompts(df=df_to_annotate, template=prompt_template, batch_size=self.batch_size)
-
-    def __call__(self, df_to_annotate: pd.DataFrame, **decoding_kwargs) -> pd.DataFrame:
-        """Annotates the given examples.
-
-        Parameters
-        ----------
-        df_to_annotate : pd.DataFrame
-            Examples to annotate
-
-        decoding_kwargs :
-            Additional arguments to pass to `fn_completions`.
-        """
-        df_to_annotate = df_to_annotate.copy()  # avoid in place modifications
-
-        if df_to_annotate.empty:
-            df_to_annotate["preference"] = []
-            return df_to_annotate
-
-        df_to_annotate = self.preprocess(df_to_annotate)
-
-        # prompts and completions here will not be the same length as the dataframe due to batching
-        prompts, df_to_annotate = self.make_prompts(df_to_annotate)
-
-        completions = self.fn_completions(prompts=prompts, **self.completions_kwargs, **decoding_kwargs)
-
-        df_to_annotate["preference"] = self.parse_completions(completions=completions["completions"])
-        for k, v in completions.items():
-            if k != "completions":
-                if len(df_to_annotate["preference"]) == len(v) * self.batch_size:
-                    v = [el for el in v for _ in range(self.batch_size)]
-                df_to_annotate[k] = v
-                if "per_example" in k:
-                    df_to_annotate[k] = df_to_annotate[k] / self.batch_size
-
-        df_annotated = self.postprocess(df_to_annotate)
-
-        return df_annotated
-
-    def preprocess(self, df_to_annotate: pd.DataFrame) -> pd.DataFrame:
-        """Preprocess the examples before annotating. In particular, takes care of all the randomization."""
-
+    def _preprocess(self, df_to_annotate: pd.DataFrame) -> pd.DataFrame:
         if self.is_randomize_output_order:
             # randomize order of output_1, output_2 base on inputs
             df_to_annotate["is_switched_outputs"] = df_to_annotate.apply(
@@ -719,41 +344,12 @@ class SinglePairwiseAnnotator:
             )
             df_to_annotate = utils.shuffle_pairwise_preferences(df_to_annotate, df_to_annotate["is_switched_outputs"])
 
-        if self.is_shuffle:
-            df_to_annotate = df_to_annotate.sample(frac=1, random_state=self.seed)
+        df_to_annotate = super()._preprocess(df_to_annotate)
 
         return df_to_annotate
 
-    def parse_completions(self, completions: list[str]) -> list[int]:
-        """Converts the completions into annotations."""
-        all_preferences = []
-        for completion in completions:
-            # use a regex to match all outputs on a line. Assumes that there is at most one output to match per line
-            batch_preferences = self.fn_completion_parser(completion)
-            if len(batch_preferences) != self.batch_size:
-                logging.warning(
-                    f"Found {len(batch_preferences)} preferences in:'''\n{completion}\n''' but expected"
-                    f" {self.batch_size}. We are setting all preferences to np.nan."
-                )
-                batch_preferences = [np.nan] * self.batch_size
-            all_preferences += batch_preferences
-        return all_preferences
-
-    def postprocess(self, df_annotated: pd.DataFrame) -> pd.DataFrame:
-        """Postprocess the annotated examples."""
-
-        # remove padding examples when using batch_size > 1
-        df_annotated = df_annotated.query("is_padding == False").drop(columns=["is_padding"])
-
-        arr_is_na = df_annotated["preference"].isna()
-        if arr_is_na.any():
-            logging.warning(
-                f"{arr_is_na.sum().item()} samples had no auto annotation. We are filtering them for now. "
-                f"If you are using chain of thought it might be that max_tokens limit is too low. "
-            )
-            df_annotated = df_annotated[~arr_is_na]
-
-        assert set(df_annotated["preference"].unique().tolist()) <= {0, 1, 2}
+    def _postprocess(self, df_annotated: pd.DataFrame) -> pd.DataFrame:
+        df_annotated = super()._postprocess(df_annotated)
 
         if self.is_randomize_output_order:
             # unshuffles output 1 and output 2. For binary preference, unshuffling is equivalent to reshuffling
